maga_transformer/requirements_torch_gpu.txt
filelock
jinja2
sympy
typing-extensions
importlib_metadata
transformers==4.33.1
sentencepiece==0.1.99
fastapi==0.95.0
uvicorn==0.21.1
dacite
pynvml
thrift
numpy==1.24.1
psutil
tiktoken==0.4.0
lru-dict
py-spy
safetensors
cpm_kernels
pyodps
Pillow
protobuf==3.20.0
matplotlib
https://download.pytorch.org/whl/cu118/torch-2.1.0%2Bcu118-cp310-cp310-linux_x86_64.whl#sha256=a81b554184492005543ddc32e96469f9369d778dedd195d73bda9bed407d6589
torchvision==0.16.0
einops
